requirements.txt
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.12
idna==3.3
joblib==1.1.0
mysql-connector==2.2.9
numpy==1.22.2
requests==2.27.1
scikit-learn==1.0.2
scipy==1.8.0
sklearn==0.0
soupsieve==2.3.1
threadpoolctl==3.1.0
urllib3==1.26.8
